router product

+++ db/models/product.py
@@ -5,7 +5,7 @@ from typing import Optional
 class Product(BaseModel):
     id: Optional[str] = None
     name: str
-    price: int
+    price: float
     detail: str
     car: bool
